The Chef can now use the reports page.
So I accidentally committed on master like an idiot.
Note to self: Dont ever hit git reset --HARD^ on master ever.
I used git reset 'HEAD@{1}' to solve but still a pain in the ass.

The Chef can select a cuisine and view results about the meals associated
to that cuisine. THe chef can also view ingredients associated to each
meal associated with the cuisine.